Move serial traits impl. back to uart.rs
