+++ rp2040-hal/src/lib.rs
@@ -23,5 +23,4 @@ pub mod ssi;
 pub mod timer;
 pub mod uart;
 pub mod usb;
-pub mod serial;
 pub mod watchdog;
